API Collection Added

+++ apis/views.py
@@ -1,3 +1,6 @@
+from rest_framework.authtoken.models import Token
+from django.contrib.auth.models import User
+from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
@@ -35,16 +38,19 @@ class CategoryDetailAPIView(APIView):
 
     def put(self, request, pk):
         category = Category.objects.get(pk=pk)
-        serializer = CategorySerializer(category, data=request.data)
+        serializer = CategorySerializer(category, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
-        category = Category.objects.get(pk=pk)
-        category.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        category = Category.objects.filter(pk=pk).first()
+        if category:
+            category.delete()
+            return Response(status=status.HTTP_200_OK)
+        else:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
 
 
 class ProductAPIView(APIView):
@@ -75,7 +81,7 @@ class ProductDetailAPIView(APIView):
 
     def put(self, request, pk):
         product = Product.objects.get(pk=pk)
-        serializer = ProductSerializer(product, data=request.data)
+        serializer = ProductSerializer(product, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
@@ -85,3 +91,40 @@ class ProductDetailAPIView(APIView):
         product = Product.objects.get(pk=pk)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+class CustomAuthToken(ObtainAuthToken):
+    def post(self, request, *args, **kwargs):
+        serializer = self.serializer_class(
+            data=request.data, context={'request': request})
+        serializer.is_valid(raise_exception=True)
+        user = serializer.validated_data['user']
+        token, created = Token.objects.get_or_create(user=user)
+        return Response({'token': token.key})
+
+
+class RegisterAPIView(APIView):
+    def post(self, request):
+        first_name = request.data.get('first_name')
+        last_name = request.data.get('last_name')
+        email = request.data.get('email')
+        username = request.data.get('username')
+        password = request.data.get('password')
+        
+        if not first_name or not last_name or not email or not password or not username:
+            return Response({'error': 'Please provide all the required fields: first_name, last_name, email, username, and password.'}, status=400)
+        
+        if User.objects.filter(username=username).exists():
+            return Response({'error': 'Username already exists.'}, status=400)
+
+        user = User.objects.create_user(
+            username=username,
+            password=password,
+            first_name=first_name,
+            last_name=last_name,
+            email=email
+        )
+
+        token, created = Token.objects.get_or_create(user=user)
+
+        return Response({'token': token.key})
